Added test case where there's one item in the basket

+++ CheckoutLibrary/ICheckout.cs
@@ -5,6 +5,6 @@ namespace CheckoutLibrary
     public interface ICheckout
     {
         void Scan(string item);
-        int GetTotalPrice();
+        decimal GetTotalPrice();
     }
 }
